Remove language server, add LSP conversion utilities
